Fix customer edit form showing incorrect value
The further contact dropdown field on the customer edit form expects an
integer value (value in db) but casted value (string) was given due to
how Rails handles the enum field has changed. As the result, incorrect
option is selected for this field.

This change fixes this issue and show the correct value selected.